% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/themes.R
\name{theme_qsp1}
\alias{theme_qsp1}
\title{QSP Theme 1}
\usage{
theme_qsp1(base_size = 11, base_family = "")
}
\arguments{
\item{base_size}{base font size}

\item{base_family}{base font family}
}
\value{
a ggplot theme
}
\description{
This is a ggplot theme with a white background and light grey gridlines.
}
\examples{

mtcars2 <-  dplyr::mutate(mtcars,
                      vs = factor(vs, labels = c("V-shaped", "Straight")),
                      am =factor(am, labels = c("Automatic", "Manual")),
                      cyl = factor(cyl),
                      gear = factor(gear))

library(ggplot2)
p1 <- ggplot2::ggplot(mtcars2) +
  geom_point(aes(x = wt, y = mpg, colour = gear)) +
  labs(title = "Fuel economy declines as weight increases",
       subtitle = "(1973-74)",
  caption = "Data from the 1974 Motor Trend US magazine.",
  tag = "Figure 1",
  x = "Weight (1000 lbs)",
  y = "Fuel economy (mpg)",
  colour = "Gears")

p1 + theme_qsp1()

}
